Add CrowdSec log processor for Minecraft
Added custom parser and scenario for Minecraft. The parser looks for
failed auth for the following cases:
- Online-mode players not whitelisted
- Online-mode banned players
- Offline-mode players
- Offline-mode players with invalid usernames

The scenario detects the above bad actors by looking for an IP spamming
>5 attempts within a 10s window.

Unfortunately, offline-mode players with valid usernames have the same
logs as whitelisted online-mode players with connectivity issues. It's
unlikely that good players with poor connectivity spam as often as bad
actors, so they shouldn't be impacted.